Add extra spacing and cutouts for controls and ports
The fit that I had previously was too snug, the microphone would not slide all the way down. I am not sure if there is a slight taper on the body of the mic or if the PLA just accomodated some stretch and couldn't go further, but it was definitely too tight of a fit.

I added some amount of play around the internal negative volume so that the microphone fits in better and also two cutouts. The pop filter arm cutout is extremely generous right now and I will eventually lock it down a bit. The one for knobs and the monitor jack matches the size of the side of the microphone and I think that one might look good lining the side so I might keep it that way.