Decrease the play and make cutouts more precise
.5 mm play seems like it is a nice fit, but maybe too lose. Let's try the middle between the zero it was before and the .5 mm it was now and see if it still fits.

The knobs/jack cutout got raised to hold the wall together better and the amount is just enough to cover the mic bottom rounded edge, so it makes it nicer, too.

The pop filter arm cutout was basically reduced by a half. The arm is 45 degrees I guess, because when positioned dead center in front of the arm, it works out nicely to half the front section leaving the whole side unchanged.